feat: create ProjectHeader and abstract MUI CardHeader from DisplayPreview and DisplayFull
