+++ src/components/IconWithTooltip/ProjectIconBox.tsx
@@ -3,17 +3,6 @@ import { icons } from '../../data/icons';
 import { IconWithTooltip } from './IconWithTooltip';
 
 
-const unfoldIconOptions = {
-  true: {
-    ariaLabel: 'See less information',
-    icon: icons.UnfoldLess,
-  },
-  false: {
-    ariaLabel: 'See more information',
-    icon: icons.UnfoldMore,
-  },
-};
-
 const githubIconOptions = (githubLink: string) => ({
   icon: {
     ariaLabel: 'View on GitHub',
@@ -53,18 +42,9 @@ export const ProjectIconBox = ({
 }) => {
   const { liveLink, githubLink } = projectLinks;
 
-  const handleToggleDisplay = () => {
-    toggleDisplay(index);
-  };
-
 
   return (
     <div className="display-flex flex-center center-column">
-      <IconWithTooltip
-        icon={unfoldIconOptions[fullDisplay.toString()]}
-        onClick={handleToggleDisplay}
-        placement='right'
-      />
       <IconWithTooltip {...githubIconOptions(githubLink)} placement='right'/>
       <IconWithTooltip {...externalLinkOptions(liveLink)} placement='right'/>
     </div>
